+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 * `MastoApi` now has a `reset()` method which is called by `TheAlgorithm` during resets
 * Add telemetry to `MastoApi` for request wait times (data is returned in `getCurrentState()` under the new `Api` section)
 * `QUICK_MODE` isn't always turned on when `NODE_ENV=development` and/the `FEDIALGO_DEBUG` env var is turned on; has to be enabled separately now
-* Add `BooleanFilter.isOptionEnabled()` method
+* Add `BooleanFilter.isThisSelectionEnabled()` method
+* ~~Fix bug where BooleanFilters weren't being fully reconstructed as class instances (they were just objects and so had no methods)~~
+* Export ComponentLogger class
 
 ### v0.54.12
 * Add a `TootCache` object property to each `Toot` that memoizes the results of frequently called string functions like `contentWithCard()` or `contentWithEmojis()`